pyproject.toml
[tool.poetry]
name = "daeclipse"
version = "0.0.4"
description = "A Python library for DeviantArt Eclipse functionality"
readme = "README.md"
authors = ["Kathryn DiPippo <dipippo.k@gmail.com>"]
license = "MIT"
homepage = "https://kathryndipippo.com/daeclipse/"
repository = "https://github.com/Pepper-Wood/daeclipse"
keywords = ["python", "deviantart"]

[tool.poetry.urls]
Issues = "https://github.com/Pepper-Wood/daeclipse/issues"

[tool.poetry.dependencies]
python = ">=3.7,<4.0.0"
beautifulsoup4 = "^4.7.1"
browser-cookie3 = "^0.12.1"
cli-ui = ">=0.14.1,<0.16.0"
pick = "^1.0.0"
requests = "^2.26.0"
typer = "^0.3.2"
python-dotenv = "^0.19.0"
html-to-draftjs = "^1.0.1"
tabulate = "^0.8.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pdoc = "^7.2.0"
mkdocs = "^1.2.2"
mkdocs-material = "^7.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
